[Pal/Linux-SGX] make return value of sgx_ocall_fn_t long
the return type of INLINE_SYSCALL is long (%rax), not int or %eax.

+++ Pal/src/host/Linux-SGX/ocall_types.h
@@ -17,7 +17,7 @@
  */
 #pragma pack(push, 1)
 
-typedef int (*sgx_ocall_fn_t)(void*);
+typedef long (*sgx_ocall_fn_t)(void*);
 
 enum {
     OCALL_EXIT = 0,
